change skeletor prefix
