+++ components/$SkeletorProvider/index.ts
@@ -1,0 +1,2 @@
+export * from './$SkeletorProvider';
+export * from './$SkeletorContext';
