+++ components/$Wrapper/$Wrapper.tsx
@@ -1,0 +1,48 @@
+import React from 'react';
+import { View, ViewProps } from 'react-native';
+import { getUsableStylesFromProps } from 'skeletor/helpers';
+
+type Props = $WrapperProps & $Alignment & $Spacing & $Size & ViewProps;
+
+export const $Wrapper: React.FC<Props> = ({
+  children,
+  margins,
+  paddings,
+  align,
+  justify,
+  flex,
+  maxHeight,
+  maxWidth,
+  height,
+  width,
+  minHeight,
+  minWidth,
+  style,
+  background,
+  ...rest
+}) => {
+  return (
+    <View
+      style={[
+        getUsableStylesFromProps({
+          ...margins,
+          ...paddings,
+          alignItems: align,
+          justifyContent: justify,
+          backgroundColor: background,
+          flex,
+          maxHeight,
+          maxWidth,
+          minHeight,
+          minWidth,
+          height,
+          width,
+        }),
+        style,
+      ]}
+      {...rest}
+    >
+      {children}
+    </View>
+  );
+};
