--- components/_Button/index.ts
@@ -1,1 +1,0 @@
-export * from './_Button';
