--- components/_Container/index.ts
@@ -1,1 +1,0 @@
-export * from './_Container';
