--- components/_Input/index.ts
@@ -1,1 +1,0 @@
-export * from './_Input';
