--- components/_Screen/index.ts
@@ -1,1 +1,0 @@
-export * from './_Screen';
