--- components/_SkeletorProvider/index.ts
@@ -1,2 +1,0 @@
-export * from './_SkeletorProvider';
-export * from './_SkeletorContext';
