--- components/_Text/index.ts
@@ -1,1 +1,0 @@
-export * from './_Text';
