--- components/_Wrapper/index.ts
@@ -1,1 +1,0 @@
-export * from './_Wrapper';
