+++ components/index.ts
@@ -1,8 +1,8 @@
-export * from './_Text';
-export * from './_Screen';
-export * from './_Image';
-export * from './_Input';
-export * from './_Wrapper';
-export * from './_Container';
-export * from './_SkeletorProvider';
-export * from './_Button';
+export * from './$Text';
+export * from './$Screen';
+export * from './$Image';
+export * from './$Input';
+export * from './$Wrapper';
+export * from './$Container';
+export * from './$SkeletorProvider';
+export * from './$Button';
