+++ types/$Border.d.ts
@@ -1,0 +1,4 @@
+type $Border = {
+  borderWidth?: number;
+  borderColor?: string;
+};
