+++ types/$Button.d.ts
@@ -1,0 +1,9 @@
+type $ButtonProps = {
+  loading?: boolean;
+  title: string | JSX.Element;
+  textProps?: $TextProps;
+  border?: _Border;
+  color?: string;
+  leftItem?: JSX.Element;
+  rightItem?: JSX.Element;
+};
