+++ types/$Container.d.ts
@@ -1,0 +1,5 @@
+type $ContainerProps = {
+  scrollable?: boolean;
+  showsVerticalScrollIndicator?: boolean;
+  bounces?: boolean;
+};
