+++ types/$Image.d.ts
@@ -1,0 +1,5 @@
+type $ImageProps = {
+  radius?: number;
+  height: number | undefined;
+  width: number | undefined;
+};
