+++ types/$Size.d.ts
@@ -1,0 +1,9 @@
+interface $Size {
+  width?: number | string;
+  height?: number | string;
+  minHeight?: number | string;
+  minWidth?: number | string;
+  maxHeight?: number | string;
+  maxWidth?: number | string;
+  flex?: number;
+}
