+++ types/$Spacing.d.ts
@@ -1,0 +1,20 @@
+interface $Spacing {
+  margins?: {
+    marginTop?: number;
+    marginBottom?: number;
+    marginLeft?: number;
+    marginRight?: number;
+    marginHorizontal?: number;
+    marginVertical?: number;
+    margin?: number;
+  };
+  paddings?: {
+    paddingTop?: number;
+    paddingBottom?: number;
+    paddingLeft?: number;
+    paddingRight?: number;
+    paddingHorizontal?: number;
+    paddingVertical?: number;
+    padding?: number;
+  };
+}
