--- types/_Border.d.ts
@@ -1,4 +1,0 @@
-type _Border = {
-  borderWidth?: number;
-  borderColor?: string;
-};
