--- types/_Button.d.ts
@@ -1,9 +1,0 @@
-type _ButtonProps = {
-  loading?: boolean;
-  title: string | JSX.Element;
-  textProps?: _TextProps;
-  border?: _Border;
-  color?: string;
-  leftItem?: JSX.Element;
-  rightItem?: JSX.Element;
-};
