--- types/_Container.d.ts
@@ -1,5 +1,0 @@
-type _ContainerProps = {
-  scrollable?: boolean;
-  showsVerticalScrollIndicator?: boolean;
-  bounces?: boolean;
-};
