--- types/_Image.d.ts
@@ -1,5 +1,0 @@
-type _ImageProps = {
-  radius?: number;
-  height: number | undefined;
-  width: number | undefined;
-};
