--- types/_Size.d.ts
@@ -1,9 +1,0 @@
-interface _Size {
-  width?: number | string;
-  height?: number | string;
-  minHeight?: number | string;
-  minWidth?: number | string;
-  maxHeight?: number | string;
-  maxWidth?: number | string;
-  flex?: number;
-}
