--- types/_Spacing.d.ts
@@ -1,20 +1,0 @@
-interface _Spacing {
-  margins?: {
-    marginTop?: number;
-    marginBottom?: number;
-    marginLeft?: number;
-    marginRight?: number;
-    marginHorizontal?: number;
-    marginVertical?: number;
-    margin?: number;
-  };
-  paddings?: {
-    paddingTop?: number;
-    paddingBottom?: number;
-    paddingLeft?: number;
-    paddingRight?: number;
-    paddingHorizontal?: number;
-    paddingVertical?: number;
-    padding?: number;
-  };
-}
