--- types/_Wrapper.d.ts
@@ -1,3 +1,0 @@
-type _WrapperProps = {
-  background?: string;
-};
